refactor(server): extract duplicate handlers to shared functions
Eliminated ~400 lines of duplication between server.go and multitenant.go
by extracting handler logic to pkg/server/handlers.go. Both single-tenant
and multi-tenant servers now call the same handler implementations.

Changes:
- Created handlers.go with shared handler functions
- Refactored server.go to delegate to shared handlers
- Refactored multitenant.go to delegate to shared handlers
- Removed duplicate implementations of:
  * saveEvent/loadEvents
  * batchEvents/streamEvents
  * position/subscriptions handlers
- Cleaned up unused imports in both files

All tests pass. No functional changes.
